requirements.txt
Flask
PyMuPDF
google-cloud-aiplatform
python-docx
python-dotenv
spacy
weasyprint
markdown2
language-tool-python
flasgger
